feat: new arg -i --input for reading urls from a file
