+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "mediafire_rs"
 description = "Downloads files and folders from mediafire"
 authors = ["NicKoehler <grillinicola@proton.me"]
-version = "0.2.1"
+version = "0.2.2"
 edition = "2024"
 license = "MIT"
 readme = "README.md"
