+++ src/config.rs
@@ -3,6 +3,7 @@ use std::path::PathBuf;
 #[derive(Debug)]
 pub struct Config {
     pub urls: Vec<String>,
+    pub input_file: Option<PathBuf>,
     pub output_path: PathBuf,
     pub max: u64,
     pub tries: u64,
@@ -14,6 +15,7 @@ pub struct Config {
 impl Config {
     pub fn new(
         urls: Vec<String>,
+        input_file: Option<PathBuf>,
         output_path: PathBuf,
         max: u64,
         tries: u64,
@@ -23,6 +25,7 @@ impl Config {
     ) -> Self {
         Self {
             urls,
+            input_file,
             output_path,
             max,
             tries,
